Added examples why the folder can be locked
When initially configuring DM Configs, it might be that the Skyline DataMiner folder is locked, in the past this was typically due to an Anti-virus program and was quite rare. However we've recently seen that NATS locks the folder as well, even if DataMiner is stopped. NATS is used by Dataminer since v10.0, so this is likely to become a more common issue.