Update The_Bookings_module.md
The color of a booking is determined by booking property "Visualbackground". A booking timeline will use that color on a booking item by default. A booking listview can show this color by adding the property column "Visualbackground" and choose column type  'Color' to show the color instead of  a color textrepresentation.