Update SoftLaunchOptions.md
We should try to avoid "mini-TOCs" that need to be manually maintained when possible. So if it's possible to have a summary text with a few links instead on a parent page, that's the way to go. This because the manual maintenance of the mini-TOCs is time-consuming and prone to errors.