Update Sales_and_Marketing.md
Small updates & tweaks.  

Note also that the DataMiner DevOps Professional program that will launch soon.  And at that point please use this pull request also to update with a link on the reference to that program in this section.

Note that there is a reference to "corporate materials" that can be used.  This can be linked to https://skyline.be/corporate-images, although I would recommend that you first contact marketing to verify if they want to retain that page on the corporate website, or if they might consider having a Dojo page where they will add more materials (pptx template, desktop backdrops, Teams backdrops, etc.).  Please check this and update with link if possible.

Can you also please add a new page with Deal Registration and link that to "register deals".  I'll provide the content for that page then.