Update Service_Delivery.md
If you can also take note to link the DataMiner DevOps Professional reference on this page once the DataMiner DevOps Professional program is launched, along with the proper documentation / information.

The CI/CD reference in here needs to be removed, because the documentation it is referring to is not meant to be there on docs (i.e. this is the internal pipeline) - the reference here is meant to be towards a new CI/CD related set of documentation that probably still needs to be written up (explained this to Simon Raine, please get in touch with him to synch on this).