Dynamic drop-down read parameter type string
The read parameter of which the write parameter has a dynamic drop-down list should be a normal string type parameter. It's the write parameter that provides the drop-down list that the user can select items from. The read parameter is simply displaying this as a string value, there is no need to link that to the dependencyId parameter. Also added the trending="false" attribute for consistency as there is no alarm monitoring present.