Update Working_with_trend_predictions.md
Some incorrect information was fixed. For instance, the old version mentioned "trend predictions can be disabled in the user settings", this is not true.
Reference was added to Proactive cap detection, a strongly related feature which has a dependency on the Trend Prediction feature.
Clarification was added on the different prediction types: users often wonder what these represent or what the difference is between the different types.
Clarification was added on why some metrics might not have a prediction.