Update UIComponentsTreeControlAlarmBubbleUp.md
The bubble up feature has unexpected effect when added or deleted from a tree control. This unexpected effect isn't described anywhere. The behavior is that the bubble up only happens when a new alarm is created. This is because then the additional property is added to the alarm. This can cause confusion when developing and confusion for users.