Update Adding_time_scoped_related_parameters_to_a_trend_graph.md
I changed the first paragraph: it incorrectly referred the lightbulb icon on the top right of the trend graph, whereas a new lightbulb icon was introduced specifically for this feature. I also added some more info in that paragraph.

In the second paragraph, I removed "a" from "in a case".

Finally, the second item in the bottom section was reformulated.