Update Placeholders_for_variables_in_shape_data_values.md
I personally think the example that is now in the documentation is already very advanced (and difficult to understand) if you are not already familiar with regular expressions. I think by adding a more simple example (by removing the Frames suffix) it is more clear what you can do with it. This suggestion originates from a custom training I provided to TV2 about advanced vision, where they need this RegExReplace, but did not hear yet of Regular Expressions. 

It also relates to this Dojo Question: https://community.dataminer.services/question/how-can-the-sum-of-parameters-with-units-be-calculated-in-a-visio/