Update Migrating_the_general_database_to_Cassandra.md
Added information on setting Heap size on Windows.
Ideally a separate page should be made "Installing a Cassandra database on a DMA" which outlines that Cassandra is installed with DataMiner and mentions peculiarities about the configuration that we ship with. This would then have a better place there. However, that is tricky to do in the short term.