Update Advanced_analytics_features_in_the_Alarm_Console.md
We are about to add a lightbulb in the alarm console. When certain features are disabled, the lightbulb will show a link to this page to inform the user on our other functionalities. This page contained all the information we needed, except for the fact that pattern matching requires elastic. I have tried to incorporate that information here.