Review
A cross-reference has been added for each user permission. The advantage of this is that when there is a change to the user permission on the DataMiner user permissions page, the cross-reference will no longer work and we'll get a warning, so that we know to update the user permission here as well.