Removed tab layout in hello world procedure
We should be very careful about when we use tab layout. Usually, we only apply this in cases where only one of the tabs is relevant to the user (e.g. in case a procedure is different depending on the DataMiner version), as a handy way to show the user only the information that is relevant for them. In other cases, I think we should only use this layout if the added value of using it weighs up against the possibility that tabs go overlooked. As such, I'd prefer to use no tab layout in the Hello world tutorial — regular headers work just as well here.