Added new overload methods
Added the new overloaded methods that were introduced by RN38153 to QActionTable and SLProtocol. Corrected the names of the arguments of SLProtocol to match with the real interface. Even though "primaryKey" sound more clear than "row", the name of the argument is "row" and if a developer uses named argument "primaryKey" in a QAction they will think it is correct from the DevPacks (generated based on these files) but when trying to run it on a DataMiner it will fail because that one will not know what is meant with "primary key" and only knows "row".