Update MOP_Migrating_a_Failover_pair_to_new_hardware.md
The procedure assumes users are migrating Failover pair to an existing cluster. It was not clear and a user with a single failover pair followed these procedure and tried to remove his agent from the cluster when it was not necessary. 
I updated the title and description based on steps: 2,6 and 7.