Added warning about possible deadlocks with DOM actions
When you configure a DOM action that triggers and automation script, there is a chance that deadlocks occur in SLAutomation if you trigger the DOM actions via another script. This is because the DOM action trigger request is a SLNet API call, which uses one of the 10 available slots of the SLAutomation connection to SLNet. As long as the DOM action script is running, that slot is taken up. If the DOM action script then sends an SLNet request as well, another slot is needed. If all 10 slots have been taken up by DOM action trigger requests, the action scripts themselves will have to wait forever, as there won't be a slot freed anymore.