Added important note about DOM editor versions
With this note, I want to warn about the potential risk where DOM configuration could break if you use a DOM editor version that is a lot older than your DataMiner agent version. It is always recommended to update the script to the version that either exactly matches the DataMiner version, or the most recent available right before that version. (Never after, since that may include features not available on the DMA)