Update KI_LDAP_users_unable_to_log_in_after_upgrade.md
Adjusting the versions and advocating for a blanket skip of the first RN which was intended to fix this issue. As the chance of it re-occurring on said version after an edit of the LDAP server is realistic.
So we'd prefer to guide customers to the version with the full fix in place.