Only elements with SNMP traps on clusters without trap forwarding are not swarmable.
Elements that can receive traps will be blocked for swarming only in case SNMP trap distribution is not enabled on all agents in the cluster. If SNMP trap distribution is enabled on the whole DMS, this limitation is not in place. See https://intranet.skyline.be/DataMiner/Lists/Release%20Notes/DispForm.aspx?ID=41356