Update UIComponentsTableForeignKeys.md to mention the possibility of multiple columns having foreign keys to the same table
I'm not sure if this is completely clear on how I wrote it, but I learned recently while working on a DVE that you can have multiple columns in the same table which reference the same foreign key. In our case this is useful since we have a "Source Network" and a "Destination Network" which both reference the same "Network" table. Let me know if I can adjust anything to explain this idea better.