Added page concerning connection string
Connection strings have some howto pages at present, but no pages explaining the context.
Because of that, and for other reasons as well we find that some users have no idea what they should or should not do.
This page is meant to guide users to a better setup and explain why their existing setup may have a specific connection string.