Update ZiineDemoSystem.md
First update on how to access Ziine (following the actions related to the Agile Workstreams). Product Marketing will record a new intro video. As the video on YouTube is outdated following the recent updates on ziine.skyline.be, I've asked marketing to unpublish the video. As soon as we have a new video, we can add it here. 

In a later stage we will add also more information on what can be found on ziine.