Replace 'dataminer.services users' with 'Cloud users'
'dataminer.services users' is confusing. These are not users on dataminer.services, instead they are automatically generated  DataMiner users that are used for features of dataminer.services (Live Sharing)

Cube also refers to these users as 'Cloud users' so it makes sense to refer to them as such in the docs as well.